apps/api/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Queue
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
httpx==0.26.0
fake-useragent==1.4.0

# AI & ML
openai==1.10.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
lightgbm==4.3.0

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
aiohttp==3.9.1

# Discord
discord-webhook==1.3.0